+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Jeong 2023 rabeprazole ([doi:10.3390/biomedicines11113021](https://doi.org/10.3390/biomedicines11113021)) -- healthy Korean adults of both genders.
+- Add Vollmer 2023 cefepime-enmetazobactam ([doi:10.1093/ofid/ofad500.2147](https://doi.org/10.1093/ofid/ofad500.2147)) -- adults with complicated urinary tract infection or acute pyelonephritis, and healthy volunteers.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.